Задание #15 Еще больше модулей
Разбила upload.js по формам (upload-form, resize-form, filter-form), но
мне что-то это не нравится, слишком много они используют друг из друга
