+++ src/main.js
@@ -1,6 +1,5 @@
 'use strict';
 
-require('./resizer');
-require('./upload');
 require('./pictures');
+require('./filter-form');
